More back navigation in code
Backstack doesn't entirely work with multiple fragment views in a single activity,
so the super.onBackPressed just closed the app prematurely. Since the app starts on
the news view by default, it's also the last view that should be visible before
the app closed due to repeated back presses.

Intents in the Main Activity now allow back navigation to specific screens, which is
always interesting