Removed unneccesary if-statement
This if-statement was meant to block the SwitchTabEvent to overwrite the currentScreen
property of MainActivity if no tabs were visible at that time. However, a better way
to manage this is by just not opening the tab in the first place if it should not have
been opened. Since this is now handled properly in MainActivity, the if-statement in
this commit could be removed safely