Updated two intent services to JobIntentServices, to work with android Oreo
On Android Oreo, a broadcast receiver no longer is allowed to start intent services.
By converting the intent services to job services, the job is automatically scheduled
on Oreo+ devices, will maintaining old behavior on pre-oreo devices.

Context passed around from the notification receiver is from now on the context
of the broadcast service itself, rather than the app context, which may be null
when a notification is received