Bugfix: Always use the right comment to build the view holder
the onBindViewHolder function in the adapter can be called with
the wrong adapter position. In this case, replies to the wrong comment
would be added to the current viewholder. Android recovers quickly, by
calling the onBindViewHolder funtion again with the correct position,
but items were never removed.

The same went for the announcement indicator, which was set to visible
if the comment is an announcement, but never set to invisible if it was
not.

This commit assumes the view for the view holder should be built from
scratch, explicitly setting all properties assures the proper comment,
with all proper values, is displayed