Subscribing to and exporting events works again on Android Oreo +
This is very tricky with how some services have to be intent services
and other services have to be scheduled as jobs, while their behavior
is very similar. Furthermore, context is not always available in
NotificationReceiver, if functions in that domain are required by
other parts of the system that create it as a class object, rather
than start it as a service (which requires an intent).

Anyway, it works again.