Fixed index-out-of-bounds error on reloading while loading more
An index out of bounds error could occur if a spinner is added to
the end of the list and then the user reloads the list. The stored
index of the spinner then no longer exists in the list, but neither
does the spanner. This edge case is now properly handled

No items text added to poll and quotes overview