Fixed issues with notifications and firebase messaging
Google deprecated some firebase stuff. Token refresh is now handled
in the notification receiver, because that's what google wants.

Every time a user logs in (which should not be too often), the
firebase token is invalidated, so a new token is generated. This ensures
every logged-in user can succesfully send a firebase token to the API (which
requires being logged in)