Only allow subscribe from notification where possible.
When subscribing from the notification is added as an option to an agenda
notification, the event is first downloaded to the background, to check
if subscribing is possible without opening the interface.

This is only the case if no registration question needs to be answered,
and the event is not a pre-published event (i.e. the event is published,
but subscribing won't be allowed until at a later time)