First start towards single model for agenda
A single model of the agenda item should be kept in the activity,
which all fragments use to populate the state. This allows updating
toolbar buttons, along with all fragments simultaniously, and is
better in line with what a user would expect

A first start towards this idea has been made in this commit. There
still are some unresolved issues, which need to be tackled.