CKI logo added to Roboto-Regular font.
The Roboto-Regular font was incorrectly specified, so never loaded
(this already caused a warning on startup since roughly 2016). This
is now fixed.

The new ttf file contains the USCKI logo (without head) at the \\u01de
character. The advantage of this is multifold:
	1. Easier processing for parser, as no Spans have to be used.
	2. Can now be used in String resources without a BBTextView
	3. Better alignment with other characters.

Especially number 3 bothered me. It looked horrid because the size
and kerning was all wrong. This is now fixed.

Number 2 fixes an issue where only the first occurence of "CKI" was
replaced in a string on parsing. Fixing would have required a loop,
because spans had to be added to the replaced character to specify
the alternative font everywhere. Now, a simple replaceAll does the
trick.