Disallow opening new photo if other transition is in progress
The media browser is, for some reason, slightly laggy in opening a
photo in full screen. When a user gets impatient and taps the same image
again or another image, the detail media view could be opened twice, breaking
all transitions, and leaving an empty white square where the target photo used
to be.

By disallowing a detail view to be opened twice this issue is solved
(fingers crossed it did not create another)