Forum topic can now be marked as read manually or automatically.
Automatic marking as read only happens if the first to last unread
posts subsequently have been fully in view for at leats 1.5 seconds.

I created this because I thought the API allowed setting mark read on
a per post basis, but this can only be done for the whole topic. But
I was not going to throw away that work :)