Finished mentor tree.
Previously, we did not, because Peter Maatman proved difficult.
He has 6 parents, two of which are Tjeu, two of the others of which
are Kambiz. He is a child of those two (four?) multiple times himself.
I don't care.

On the USCKI website, more information about mentor groups is available,
such as an introductory text, and the starting year. This is not in the
API, so left for future work.